--- lib/features/auth/pages/login_page.dart
@@ -1,123 +1,0 @@
-import 'package:email_validator/email_validator.dart';
-import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
-import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:tokyo_hakkason2024_app/features/auth/pages/signup_page.dart';
-import 'package:tokyo_hakkason2024_app/features/home/pages/home_page.dart';
-
-import '../../../core/providers/supabase_provider.dart';
-
-class LoginPage extends StatefulWidget {
-  const LoginPage({super.key});
-  static const String routeName = '/login';
-
-  @override
-  State<LoginPage> createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
-  final _focusNode = FocusNode();
-  final _formKey = GlobalKey<FormState>();
-  bool _isLoading = false;
-
-  Future<void> _signIn() async {
-    try {
-      _focusNode.unfocus();
-      final validate = _formKey.currentState?.validate();
-      if (validate == false) {
-        _focusNode.requestFocus();
-        return;
-      }
-      setState(() => _isLoading = true);
-
-      await supabase.auth.signInWithPassword(
-        email: _emailController.text.trim(),
-        password: _passwordController.text.trim(),
-      );
-
-      if (mounted) {
-        context.go(HomePage.routeName);
-      }
-    } on AuthException catch (e) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text(e.message)),
-      );
-    } catch (e) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('予期せぬエラーが発生しました')),
-      );
-    } finally {
-      if (mounted) setState(() => _isLoading = false);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Focus(
-      focusNode: _focusNode,
-      child: Scaffold(
-        appBar: AppBar(title: const Text('ログイン')),
-        body: Padding(
-          padding: const EdgeInsets.all(16),
-          child: Form(
-            key: _formKey,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                TextFormField(
-                  controller: _emailController,
-                  decoration: const InputDecoration(labelText: 'メールアドレス'),
-                  keyboardType: TextInputType.emailAddress,
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'メールアドレスが空です';
-                    }
-                    if (!EmailValidator.validate(value)) {
-                      return 'メールアドレスが不正です';
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 16),
-                // TODO(Minato): パスワードの表示非表示を追加したい
-                TextFormField(
-                  controller: _passwordController,
-                  decoration: const InputDecoration(labelText: 'パスワード'),
-                  obscureText: true,
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'パスワードが空です';
-                    }
-                    return null;
-                  },
-                ),
-                TextButton(
-                    onPressed: () {
-                      _focusNode.unfocus();
-                      context.go(SignupPage.routeName);
-                    },
-                    child: const Text('新規登録')),
-                const SizedBox(height: 24),
-                ElevatedButton(
-                  onPressed: _isLoading ? null : _signIn,
-                  child: _isLoading
-                      ? const CircularProgressIndicator()
-                      : const Text('ログイン'),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    super.dispose();
-  }
-}
