+++ lib/router.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:tokyo_hakkason2024_app/core/providers/supabase_provider.dart';
-import 'package:tokyo_hakkason2024_app/features/auth/pages/login_page.dart';
+import 'package:tokyo_hakkason2024_app/features/auth/pages/registration_page.dart';
 import 'package:tokyo_hakkason2024_app/features/auth/pages/signup_page.dart';
 import 'package:tokyo_hakkason2024_app/features/home/pages/home_page.dart';
 
@@ -12,7 +12,7 @@ final routerProvider = Provider<GoRouter>((ref) {
     redirect: (_, state) {
       if (supabase.auth.currentUser == null &&
           state.matchedLocation == HomePage.routeName) {
-        return LoginPage.routeName;
+        return RegistrationPage.routeName;
       }
       return null;
     },
@@ -28,12 +28,12 @@ final routerProvider = Provider<GoRouter>((ref) {
         },
       ),
       GoRoute(
-        path: LoginPage.routeName,
-        name: LoginPage.routeName,
+        path: RegistrationPage.routeName,
+        name: RegistrationPage.routeName,
         pageBuilder: (context, state) {
           return MaterialPage(
             key: state.pageKey,
-            child: const LoginPage(),
+            child: const RegistrationPage(),
           );
         },
       ),
